refactor Onyx staking demo app components
