+++ packages/launcher/static/applications/onyx-stats/src/Components/ui-kit/Colors.js
@@ -1,0 +1,10 @@
+export default {
+  lightBlueColor: '#00A7E7',
+  darkBlueColor: '#131f38',
+  extraDarkBlue: '#0e1728',
+  darkGreyBlueColor: '#1d2b47',
+  medGreyBlueColor: '#57647f',
+  redColor: '#c8325a',
+  lightGreyBlue: '#b3bbcc',
+  whiteColor: '#ffffff',
+}
