+++ packages/launcher/static/applications/onyx-stats/src/index.js
@@ -1,6 +1,6 @@
 import { AppRegistry } from 'react-native-web'
 import './index.css'
-import App from './App'
+import App from './Components/App'
 
 // register the app
 AppRegistry.registerComponent('App', () => App)
